Add script that will route traffic through a Wireguard tunnel
The script provides a way to create (and delete) a network namespace
called physical where the physical network interfaces are assigned to.
The physical network devices are then managed within this network namespace.

All traffic from and to the root namespace is subsequently routed
through a Wireguard tunnel that will be the only network interface left
in the root namespace if not configured otherwise.